requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.2
Flask==2.3.2
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
pandas==2.0.1
pydantic==1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic-bin==0.4.14
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.1.0
typing_extensions==4.6.1
tzdata==2023.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
yarl==1.9.2
